+++ index.py
@@ -1,0 +1,77 @@
+import dash_core_components as dcc
+import dash_html_components as html
+from dash.dependencies import Input, Output, State
+import dash_bootstrap_components as dbc
+
+from app import app
+# import all pages in the app
+from apps import family_house, home, exploratory
+
+# building the navigation bar
+# https://github.com/facultyai/dash-bootstrap-components/blob/master/examples/advanced-component-usage/Navbars.py
+
+
+nav_pages = dbc.Row(
+    [
+        dbc.NavItem(dbc.NavLink("Home", href="/home")),
+        dbc.NavItem(dbc.NavLink("Exploratory", href="/exploratory")),
+        dbc.NavItem(dbc.NavLink("family-house", href="/family_house")),
+    ],
+    className="g-0 ms-auto flex-nowrap mt-3 mt-md-0",
+    align="center",
+)
+navbar = dbc.Navbar(
+    dbc.Container(
+        [
+            html.A(
+                # Use row and col to control vertical alignment of logo / brand
+                dbc.Row(
+                    [
+                        # dbc.Col(html.Img(src="/assets/virus.png", height="30px")),
+                        dbc.Col(dbc.NavbarBrand("Harry Potter DASH", className="ml-2")),
+                    ],
+                    align="center",
+                ),
+                href="/home",
+            ),
+            dbc.NavbarToggler(id="navbar-toggler2"),
+            nav_pages,
+        ]
+    ),
+    color="dark",
+    dark=True,
+    className="mb-4",
+)
+
+def toggle_navbar_collapse(n, is_open):
+    if n:
+        return not is_open
+    return is_open
+
+for i in [2]:
+    app.callback(
+        Output(f"navbar-collapse{i}", "is_open"),
+        [Input(f"navbar-toggler{i}", "n_clicks")],
+        [State(f"navbar-collapse{i}", "is_open")],
+    )(toggle_navbar_collapse)
+
+# embedding the navigation bar
+app.layout = html.Div([
+    dcc.Location(id='url', refresh=False),
+    navbar,
+    html.Div(id='page-content')
+])
+
+
+@app.callback(Output('page-content', 'children'),
+              [Input('url', 'pathname')])
+def display_page(pathname):
+    if pathname == '/family_house':
+        return family_house.layout
+    elif pathname == '/exploratory':
+        return exploratory.layout
+    else:
+        return home.layout
+
+if __name__ == '__main__':
+    app.run_server(debug=True)
